Maak gebruik van model in MaaltijdRepetitieForm

+++ sql/1611_maalcie.sql
@@ -10,6 +10,6 @@ ALTER TABLE mlt_aanmeldingen CHANGE aantal_gasten aantal_gasten int(11) NOT NULL
 ALTER TABLE mlt_aanmeldingen CHANGE gasten_eetwens gasten_eetwens varchar(255) NULL DEFAULT NULL;
 
 
-ALTER TABLE mlt_repetities CHANGE dag_vd_week dag_vd_week tinyint(1) NOT NULL;
+ALTER TABLE mlt_repetities CHANGE dag_vd_week dag_vd_week int(11) NOT NULL;
 ALTER TABLE mlt_repetities CHANGE standaard_prijs standaard_prijs int(11) NOT NULL;
 ALTER TABLE mlt_repetities CHANGE abonneerbaar abonneerbaar tinyint(1) NOT NULL;
